+++ extensions/word-search/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Updated dependencies to latest versions
 - Rewrote API using `useFetch` hook
 - Added option to use selection as search text ([#11748](https://github.com/raycast/extensions/issues/11748))
+- Added search for words that sound like the search text
 
 ## [Output capitalized word] - 2023-02-28